analyze: fix SoftRebootsCount property signature
The SoftRebootsCount entry in the boot-times map declared signature "t" but the
manager exports it as "u". The mismatch was harmless while the declared signature
was never checked, but becomes a dropped property once it is enforced. Declare "u".

Follow-up for 259046cfe8de29c44070b596fe9e0471e931cbc4

(cherry picked from commit 7cc0f65d63b201cc6cab71dbb1c6fcb7ce98bbbe)
(cherry picked from commit d35d1c538b834feb8b8359893ac3a64ede0c2d6f)
(cherry picked from commit bf1da821949cc55794b8fadcdf6b6694953f07d7)

+++ src/analyze/analyze-time-data.h
@@ -26,7 +26,7 @@ typedef struct BootTimes {
         usec_t initrd_unitsload_start_time;
         usec_t initrd_unitsload_finish_time;
         /* Not strictly a timestamp, but we are going to show it next to the other timestamps */
-        uint64_t soft_reboots_count;
+        uint32_t soft_reboots_count;
 
         /*
          * If we're analyzing the user instance, all timestamps will be offset by its own start-up timestamp,
